refactor: fsd 적용 - 파티룸, 프로필 관련 일부
- components/features/parties
- components/features/profile
- api/react-query/parties
- api/react-query/user
