--- src/api/react-query/parties/keys.ts
@@ -1,1 +1,0 @@
-export const PARTY_ROOMS_QUERY_KEY = 'party-rooms';
