--- src/api/react-query/user/keys.ts
@@ -1,1 +1,0 @@
-export const USER_PROFILE_QUERY_KEY = 'user/profile';
