+++ src/entities/avatar/lib/avatar.store.ts
