+++ src/entities/profile/index.ts
@@ -1,1 +1,2 @@
-export { useProfileStore } from './model/profile.model';
+export { useProfileStore } from './lib/profile.store';
+export { default as ProfileHydration } from './ui/hydration.component';
