+++ src/shared/api/react-query-keys.ts
@@ -1,0 +1,4 @@
+export enum QueryKeys {
+  MyProfile = 'my-profile',
+  PartyRoomList = 'party-room/list',
+}
