+++ src/widgets/sidebar/index.ts
@@ -1,0 +1,1 @@
+export { default as Sidebar } from './ui/sidebar.component';
